requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
asyncpg==0.29.0
APScheduler==3.10.4
google-generativeai==0.5.4